+++ source/features/locked-issue.tsx
@@ -1,61 +1,10 @@
 import React from 'react';
-import elementReady from 'element-ready';
 import LockIcon from 'octicons-plain-react/Lock';
 import * as pageDetect from 'github-url-detection';
 
 import features from '../feature-manager.js';
 import observe from '../helpers/selector-observer.js';
-import {hasToken} from '../options-storage.js';
-import api from '../github-helpers/api.js';
-import GetIssueLockStatus from './locked-issue.gql';
-import {getConversationNumber} from '../github-helpers/index.js';
-import {isInitialLoad} from '../helpers/feature-helpers.js';
-
-async function isConversationLockedViaApi(): Promise<boolean | undefined> {
-	if (!hasToken()) {
-		return undefined;
-	}
-
-	const {repository} = await api.v4uncached(GetIssueLockStatus, {
-		variables: {
-			number: getConversationNumber()!,
-		},
-	});
-
-	return repository.issueOrPullRequest.locked;
-}
-
-async function isConversationLockedViaDom(): Promise<boolean | undefined> {
-	// Only use signals that clearly indicate the lock state
-	// The form only appears to moderators
-	const lockToggle = await elementReady('.discussion-sidebar-item svg.octicon-key + strong');
-	return lockToggle ? lockToggle.textContent === 'Unlock conversation' : undefined;
-}
-
-async function isConversationLockedViaReactData(): Promise<boolean | undefined> {
-	if (!isInitialLoad()) {
-		return;
-	}
-
-	const data = await elementReady('[data-target="react-app.embeddedData"]');
-	return data ? JSON.parse(data.textContent).payload?.preloadedQueries[0].result.data.repository?.issue?.locked : undefined;
-}
-
-async function isConversationLocked(): Promise<boolean | undefined> {
-	// Like Promise.race, but it only resolves if the result is not undefined
-	return new Promise(resolve => {
-		// TODO: Add AbortSignal after https://github.com/sindresorhus/element-ready/issues/45
-		const resolveIfDefined = async (check: () => Promise<boolean | undefined>): Promise<void> => {
-			const result = await check();
-			if (result !== undefined) {
-				resolve(result);
-			}
-		};
-		resolveIfDefined(isConversationLockedViaReactData);
-		resolveIfDefined(isConversationLockedViaDom);
-		resolveIfDefined(isConversationLockedViaApi);
-	});
-}
+import isConversationLocked from '../github-helpers/is-conversation-locked.js';
 
 function LockedIndicator(): JSX.Element {
 	return (
